# packages/api-server/src/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: ConfigBuddy CMDB API
  version: 2.0.0
  description: |
    Comprehensive REST API for ConfigBuddy - Enterprise Configuration Management Database (CMDB) platform.

    ConfigBuddy v2.0 features:
    - Graph-based CI relationships with Neo4j
    - 37 connectors (17 TypeScript + 20 JSON-only)
    - Unified credential system with affinity matching
    - Discovery agent architecture with smart routing
    - AI-powered anomaly detection and impact analysis
    - Identity resolution and reconciliation

    ## Authentication

    All API endpoints (except `/api/v1/auth/login`) require authentication via:
    - **JWT Token**: Include `Authorization: Bearer <token>` header
    - **API Key**: Include `X-API-Key: <key>` header

    ## Rate Limiting

    Rate limits apply per authentication token/API key:
    - **Standard**: 1000 requests/hour
    - **Burst**: 100 requests/minute

    Rate limit headers included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: ConfigBuddy Support
    url: http://localhost:8080
    email: support@configbuddy.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://staging-api.configbuddy.io/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /credential-sets/{id}/select:
    post:
      tags: [Credential Sets]
      summary: Select credentials
      description: Select credentials from set using strategy
      operationId: selectCredentials
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _context:
                  $ref: '#/components/schemas/CredentialMatchContext'
                _strategy:
                  type: string
                  enum: [sequential, parallel, adaptive]
      responses:
        '200':
          description: Selected credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credential'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # DISCOVERY DEFINITIONS
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /auth/api-keys endpoint
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired JWT token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: CI with ID 'srv-prod-web-01' not found
...
  schemas:
...
    Credential:
      type: object
      required: [id, name, protocol, scope, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        protocol:
          $ref: '#/components/schemas/AuthProtocol'
        scope:
          $ref: '#/components/schemas/CredentialScope'
        credentials:
          type: object
          description: Encrypted credentials (redacted in responses)
        affinity:
          $ref: '#/components/schemas/CredentialAffinity'
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        last_used_at:
          type: string
          format: date-time
...
    AuthProtocol:
      type: string
      enum:
        - oauth2
        - api_key
        - basic
        - bearer
        - aws_iam
        - azure_sp
        - gcp_sa
        - ssh_key
        - ssh_password
        - certificate
        - kerberos
        - snmp_v2c
        - snmp_v3
        - winrm
    CredentialScope:
      type: string
      enum: [cloud_provider, ssh, api, network, database, container, universal]
    CredentialAffinity:
      type: object
      properties:
        _networks:
          type: array
          items:
            type: string
          description: CIDR network ranges
          example: ["10.0.0.0/8"]
        _hostname_patterns:
          type: array
          items:
            type: string
          description: Regex patterns for hostname matching
          example: [".*\\.prod\\.example\\.com"]
        _os_types:
          type: array
          items:
            type: string
          example: [ubuntu, debian, centos]
        _device_types:
          type: array
          items:
            type: string
          example: [server, virtual-machine]
        _environments:
          type: array
          items:
            type: string
          example: [production, staging]
        _cloud_providers:
          type: array
          items:
            type: string
          example: [aws, azure, gcp]
        _priority:
          type: integer
          minimum: 1
          maximum: 10
          description: Priority when multiple credentials match (higher = preferred)
    CredentialMatchContext:
      type: object
      properties:
        _ip:
          type: string
        _hostname:
          type: string
        _os_type:
          type: string
        _device_type:
          type: string
        _environment:
          type: string
        _cloud_provider:
          type: string
        _required_protocol:
          $ref: '#/components/schemas/AuthProtocol'
        _required_scope:
          $ref: '#/components/schemas/CredentialScope'
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type/code
          example: Validation Error
        message:
          type: string
          description: Human-readable error message
          example: Invalid CI type provided
        details:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
              error:
                type: string