# packages/api-server/src/rest/controllers/credential-set.controller.ts
import { Request, Response } from 'express';
import { getCredentialSetService, getPostgresClient } from '@cmdb/database';
import {
  CredentialSetInput,
  CredentialSetUpdateInput,
  CredentialMatchContext,
  CredentialSetStrategy,
  logger,
} from '@cmdb/common';

/**
 * Credential Set Controller
 * Handles REST API requests for managing credential sets
 */
export class CredentialSetController {
  private credentialSetService;

  constructor() {
    const pool = getPostgresClient().pool;
    this.credentialSetService = getCredentialSetService(pool);
  }

  /**
   * POST /api/v1/credential-sets - Create credential set
   */
  async create(req: Request, res: Response): Promise<void> {
    try {
      const input: CredentialSetInput = req.body;
      const createdBy = (req as any).user?.id || 'system'; // Get from auth middleware

      const credentialSet = await this.credentialSetService.create(input, createdBy);

      res.status(201).json({
        _success: true,
        _data: credentialSet,
        _message: 'Credential set created successfully',
      });
    } catch (error) {
      logger.error('Error creating credential set', { error });

      // Check for duplicate name error
      if (error instanceof Error && error.message.includes('already exists')) {
        res.status(409).json({
          _success: false,
          _error: 'Conflict',
          _message: error.message,
        });
        return;
      }

      // Check for invalid credential IDs error
      if (error instanceof Error && error.message.includes('do not exist')) {
        res.status(400).json({
          _success: false,
          _error: 'Bad Request',
          _message: error.message,
        });
        return;
      }

      res.status(500).json({
        _success: false,
        _error: 'Failed to create credential set',
        _message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }

  /**
   * GET /api/v1/credential-sets - List all credential sets with expanded credentials
   */
  async list(_req: Request, res: Response): Promise<void> {
    try {
      const credentialSets = await this.credentialSetService.list();

      res.status(200).json({
        _success: true,
        _data: credentialSets,
        _count: credentialSets.length,
      });
    } catch (error) {
      logger.error('Error listing credential sets', { error });
      res.status(500).json({
        _success: false,
        _error: 'Failed to list credential sets',
        _message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }

  /**
   * GET /api/v1/credential-sets/:id - Get credential set by ID
   */
  async getById(req: Request, res: Response): Promise<void> {
    try {
      const { id } = req.params;

      if (!id) {
        res.status(400).json({
          _success: false,
          _error: 'Bad Request',
          _message: 'Credential set ID is required',
        });
        return;
      }

      // Get credential set with expanded credentials
      const credentialSet = await this.credentialSetService.getWithCredentials(id);

      if (!credentialSet) {
        res.status(404).json({
          _success: false,
          _error: 'Not Found',
          _message: `Credential set with ID '${id}' not found`,
        });
        return;
      }

      res.status(200).json({
        _success: true,
        _data: credentialSet,
      });
    } catch (error) {
      logger.error('Error getting credential set', { error, id: req.params['id'] });
      res.status(500).json({
        _success: false,
        _error: 'Failed to get credential set',
        _message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }

  /**
   * PUT /api/v1/credential-sets/:id - Update credential set
   */
  async update(req: Request, res: Response): Promise<void> {
    try {
      const { id } = req.params;

      if (!id) {
        res.status(400).json({
          _success: false,
          _error: 'Bad Request',
          _message: 'Credential set ID is required',
        });
        return;
      }

      const input: CredentialSetUpdateInput = req.body;

      const credentialSet = await this.credentialSetService.update(id, input);

      res.status(200).json({
        _success: true,
        _data: credentialSet,
        _message: 'Credential set updated successfully',
      });
    } catch (error) {
      logger.error('Error updating credential set', { error, id: req.params['id'] });

      // Check for not found error
      if (error instanceof Error && error.message.includes('not found')) {
        res.status(404).json({
          _success: false,
          _error: 'Not Found',
          _message: error.message,
        });
        return;
      }

      // Check for duplicate name error
      if (error instanceof Error && error.message.includes('already exists')) {
        res.status(409).json({
          _success: false,
          _error: 'Conflict',
          _message: error.message,
        });
        return;
      }

      // Check for invalid credential IDs error
      if (error instanceof Error && error.message.includes('do not exist')) {
        res.status(400).json({
          _success: false,
          _error: 'Bad Request',
          _message: error.message,
        });
        return;
      }

      res.status(500).json({
        _success: false,
        _error: 'Failed to update credential set',
        _message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }

  /**
   * DELETE /api/v1/credential-sets/:id - Delete credential set
   */
  async delete(req: Request, res: Response): Promise<void> {
    try {
      const { id } = req.params;

      if (!id) {
        res.status(400).json({
          _success: false,
          _error: 'Bad Request',
          _message: 'Credential set ID is required',
        });
        return;
      }

      await this.credentialSetService.delete(id);

      res.status(204).send();
    } catch (error) {
      logger.error('Error deleting credential set', { error, id: req.params['id'] });

      // Check for not found error
      if (error instanceof Error && error.message.includes('not found')) {
        res.status(404).json({
          _success: false,
          _error: 'Not Found',
          _message: error.message,
        });
        return;
      }

      // Check for in-use error
      if (error instanceof Error && error.message.includes('currently used')) {
        res.status(409).json({
          _success: false,
          _error: 'Conflict',
          _message: error.message,
        });
        return;
      }

      res.status(500).json({
        _success: false,
        _error: 'Failed to delete credential set',
        _message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }

  /**
   * POST /api/v1/credential-sets/:id/select - Select credentials with strategy
   * Returns ordered list of credentials based on the set's strategy and context
   */
  async select(req: Request, res: Response): Promise<void> {
    try {
      const { id } = req.params;

      if (!id) {
        res.status(400).json({
          _success: false,
          _error: 'Bad Request',
          _message: 'Credential set ID is required',
        });
        return;
      }

      const context: CredentialMatchContext = req.body.context || {};
      const strategy: CredentialSetStrategy | undefined = req.body.strategy;

      const credentials = await this.credentialSetService.selectCredentials(
        id,
        context,
        strategy
      );

      // Redact sensitive credentials before returning
      const safeCredentials = credentials.map((credential) => ({
        ...credential,
        credentials: '***REDACTED***',
      }));

      res.status(200).json({
        _success: true,
        _data: safeCredentials,
        _count: safeCredentials.length,
      });
    } catch (error) {
      logger.error('Error selecting credentials from set', { error, id: req.params['id'] });

      // Check for not found error
      if (error instanceof Error && error.message.includes('not found')) {
        res.status(404).json({
          _success: false,
          _error: 'Not Found',
          _message: error.message,
        });
        return;
      }

      res.status(500).json({
        _success: false,
        _error: 'Failed to select credentials',
        _message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }
}